Change vector set function to double Add ft_memsetd for double

+++ maths/basics/ft_trig.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-static float	cos_mod(double x)
+static double	cos_mod(double x)
 {
 	double	result;
 	double	den;
