+++ maths/vectors/ivec2_conversion.c
@@ -14,15 +14,15 @@
 
 t_vec2	iv2_to_v2(t_ivec2 vec)
 {
-	return (vec2((float)vec.x, (float)vec.y));
+	return (vec2((double)vec.x, (double)vec.y));
 }
 
 t_vec3	iv2_to_v3(t_ivec2 vec)
 {
-	return (vec3((float)vec.x, (float)vec.y, 0.0f));
+	return (vec3((double)vec.x, (double)vec.y, 0.0f));
 }
 
 t_vec4	iv2_to_v4(t_ivec2 vec)
 {
-	return (vec4((float)vec.x, (float)vec.y, 0.0f, 0.0f));
+	return (vec4((double)vec.x, (double)vec.y, 0.0f, 0.0f));
 }
