+++ maths/vectors/vec2_magnitude.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-float	vec2_mag(t_vec2 vec)
+double	vec2_mag(t_vec2 vec)
 {
 	return (ft_sqrt(ft_pow2(vec.x) + ft_pow2(vec.y)));
 }
